+++ src/main/java/cn/nukkit/inventory/ContainerInventory.java
@@ -35,7 +35,7 @@ public abstract class ContainerInventory extends BaseInventory {
         ContainerOpenPacket pk = new ContainerOpenPacket();
         pk.windowid = (byte) who.getWindowId(this);
         pk.type = this.getType().getNetworkType();
-        pk.slots = (short) this.getSize();
+        pk.slots = this.getSize();
         InventoryHolder holder = this.getHolder();
         if (holder instanceof Vector3) {
             pk.x = (int) ((Vector3) holder).x;
