+++ src/main/java/cn/nukkit/network/protocol/AddEntityPacket.java
@@ -48,7 +48,7 @@ public class AddEntityPacket extends DataPacket {
         this.putFloat(this.yaw);
         this.putFloat(this.pitch);
         this.put(Binary.writeMetadata(this.metadata));
-        this.putShort((short) this.links.length);
+        this.putShort(this.links.length);
         for (Object[] link : links) {
             this.putLong((Long) link[0]);
             this.putLong((Long) link[1]);
